adding documentation to tokenCommand
